Moved votes, catalog, and feeds in tests
Created a way to run dependencies for modules
