+++ backend/dataall/modules/feed/api/__init__.py
@@ -1,1 +1,1 @@
-from dataall.modules.feed.api import types
+from dataall.modules.feed.api import resolvers, input_types, types, queries, mutations, registry
