+++ tests/modules/datasets/test_dataset_feed.py
@@ -1,5 +1,5 @@
 
-from dataall.core.feed.api.registry import FeedRegistry
+from dataall.modules.feed.api.registry import FeedRegistry
 from dataall.modules.datasets_base.db.models import DatasetTableColumn
