+++ tests/modules/vote/test_vote.py
